Move `@babel/traverse` helpers generation to `Makefile`
