+++ packages/babel-traverse/scripts/generators/virtual-types.js
